Add unittest for the PatchWriter.
Split the unittest logic between bsdiff() function and the PatchWriter.
This patch now tests individually the PatchWriter logic that writes to
disk and the bsdiff() logic to create patches in the same very simple
cases.

We also update the Android.bp file to only pass -DBSDIFF_TARGET_UNITTEST
when building for the target.

Bug: 34220646
Test: ran unittests.

Change-Id: I9350157ee4a0b5a617f44bb187bd7652f6d6f017