bsdiff: Redo the gyp file
New changes in bsdiff causes a lot of undefined reference failures in
the CrOS side. This CL modifies the bsdiff.gyp file to create a static
and shared libraries for both bsdiff and bspatch. static libraries are
used for unittests, and shared libraries are used in executables and by
other external libraries. This way we do not have to export all the
symbols that are not required to be exported.

BUG=chromium:783437
TEST=unittests pass; 'bsdiff' and 'bspatch' run with no dependency issues.

Change-Id: If6ad895756f7ef797844ba95e3dfd174793e10a8