Modify the patch writer to use a list of compressors for each stream
Switch the internal member of the patch writer from a single compressor to a
list of compressors for ctrl|diff|extra streams. The change doesn't affect the
functionality of the existing bz2|brotli compressors; but later allow us to try
all the compressors and select the one with the best compression ratio.

With this new writer, callers will be able to produce the smallest patch without
calling bsdiff explictly multiple times.

Test: unit tests pass
Change-Id: I441ae6636e2299b2cfc9d14f50952c85585dbb4f